setup resources and initial assets
