+++ examples/project_ask_ai_dagster/.gitignore
@@ -4,3 +4,4 @@ __pycache__/
 test.py
 github_knowledge_base/
 .vscode/
+*.sqlite3
