+++ examples/project_ask_ai_dagster/requirements.txt
@@ -1,9 +1,11 @@
 langchain
+langchain-core
 gql
 python-dotenv
 langchain-community
 langchain-openai
 dagster
+dagster-webserver
 chromadb
 tenacity
 tqdm
